PDFBOX-4071: remove uneeded unboxing
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1823366 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/contentstream/PDAbstractContentStream.java
@@ -222,7 +222,7 @@ public abstract class PDAbstractContentStream implements Closeable
             }
             else if (obj instanceof Float)
             {
-                writeOperand(((Float) obj).floatValue());
+                writeOperand((Float) obj);
             }
             else
             {
